clean temp file method names

+++ lib/asperalm/temp_file_manager.rb
@@ -11,23 +11,27 @@ module Asperalm
       @created_files=[]
     end
 
-    def temp_filelist_path(temp_folder)
-      FileUtils::mkdir_p(temp_folder) unless Dir.exist?(temp_folder)
-      new_file=File.join(temp_folder,SecureRandom.uuid)
-      @created_files.push(new_file)
-      return new_file
-    end
-
+    # call this on process exit
     def cleanup
       @created_files.each do |filepath|
-        File.delete(filepath)
+        File.delete(filepath) if File.file?(filepath)
       end
       @created_files=[]
     end
 
-    def global_tmpfile_path(some_name)
-      username = Etc.getlogin || Etc.getpwuid(Process.uid).name || 'unknown' rescue 'unknown'
-      return File.join(Etc.systmpdir,username)+'_'+some_name
+    # ensure that provided folder exists, or create it, generate a unique filename
+    # @return path to that unique file
+    def new_file_path_in_folder(temp_folder,add_base='')
+      FileUtils::mkdir_p(temp_folder) unless Dir.exist?(temp_folder)
+      new_file=File.join(temp_folder,add_base+SecureRandom.uuid)
+      @created_files.push(new_file)
+      return new_file
+    end
+
+    # same as above but in global temp folder
+    def new_file_path_global(base_name)
+      username = Etc.getlogin || Etc.getpwuid(Process.uid).name || 'unknown_user' rescue 'unknown_user'
+      return new_file_path_in_folder(Etc.systmpdir,base_name+'_'+username+'_')
     end
   end
 end
